tools/mtp-release/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mtp-release"
version = "1.0.0"
description = "Release contract and provider matrix tooling for MTP"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Lubor Fedak", email = "lubor@codemoravia.com" },
]
dependencies = [
    "click>=8.0",
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.scripts]
mtp-release = "mtp_release.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mtp_release"]
